Fix a subtle timing issue on Node v7
The `finish` event is fired when a writable stream can not be written to anymore. The `end` event is fired when a readable stream can not be read any more.

When downstream of a readable stream, to tell that the stream is done you're interested in the `end` event, not the `finish` event.